% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{rpc_serialize}
\alias{rpc_serialize}
\title{Convert \R Objects into the \code{XML-RPC} Format}
\usage{
rpc_serialize(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{additional optional arguments (currently ignored).}
}
\value{
an object of class \code{"xml_node"}.
}
\description{
Serialize \R Objects so they can be passed to 
  \code{to_xmlrpc} as parameters.
}
\examples{
rpc_serialize(1L)
rpc_serialize(1:2)
rpc_serialize(LETTERS[1:2])
}
