% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlrpc2.R
\name{to_xmlrpc}
\alias{to_xmlrpc}
\title{Create a \code{XML-RPC} Call}
\usage{
to_xmlrpc(method, params)
}
\arguments{
\item{method}{a character string giving the name of the method 
to be invoked.}

\item{params}{a list containing the parmeters which are added to 
the \code{XML} file sent via the remote procedure call.}
}
\value{
an object of class \code{"xml_node"} containing a \code{XML-RPC} call.
}
\description{
abc
}
\examples{
params <- list(1L, 1:3, rnorm(3), LETTERS[1:3], charToRaw("A"))
cat(as.character(to_xmlrpc("some_method", params)))
}
